chore(#15): reorganize test_utils
