+++ src/client.rs
@@ -2,7 +2,7 @@ pub mod args;
 pub mod handlers;
 pub mod logging;
 pub mod shell;
-pub mod types;
 #[cfg(test)]
-pub mod utils;
+pub mod test_utils;
+pub mod types;
 pub mod validation;
