+++ src/client/test_utils/assertions.rs
@@ -1,0 +1,3 @@
+pub mod command;
+pub mod terrain;
+pub mod zsh;
