+++ src/client/test_utils/constants.rs
@@ -1,0 +1,27 @@
+pub const IN_CURRENT_DIR: bool = false;
+pub const IN_CENTRAL_DIR: bool = true;
+pub const WITH_EMPTY_TERRAIN_TOML: &str = "./tests/data/terrain.empty.toml";
+pub const WITH_NONE_BIOME_FOR_EMPTY_TERRAIN_SCRIPT: &str = "./tests/data/terrain-none.empty.zsh";
+pub const WITH_EXAMPLE_TERRAIN_TOML_COMMENTS: &str = "./tests/data/terrain.example.comments.toml";
+pub const WITH_EXAMPLE_TERRAIN_TOML: &str = "./tests/data/terrain.example.toml";
+pub const WITH_NONE_BIOME_FOR_EXAMPLE_SCRIPT: &str = "./tests/data/terrain-none.example.zsh";
+pub const WITH_EXAMPLE_BIOME_FOR_EXAMPLE_SCRIPT: &str =
+    "./tests/data/terrain-example_biome.example.zsh";
+pub const WITH_EXAMPLE_TERRAIN_TOML_COMMENTS_SPACES: &str =
+    "./tests/data/terrain.example.comments.spaces.toml";
+pub const WITHOUT_DEFAULT_BIOME_TOML: &str = "./tests/data/terrain.example.without.default.toml";
+pub const WITH_NEW_EXAMPLE_BIOME2_EXAMPLE_TOML: &str =
+    "./tests/data/terrain.example.new.example_biome2.toml";
+pub const WITH_EXAMPLE_BIOME2_FOR_EXAMPLE_SCRIPT: &str =
+    "./tests/data/terrain-example_biome2.example.zsh";
+pub const WITH_EXAMPLE_BIOME_UPDATED_EXAMPLE_TOML: &str =
+    "./tests/data/terrain.example.example_biome.updated.toml";
+pub const WITH_EXAMPLE_BIOME_FOR_UPDATED_EXAMPLE_BIOME_SCRIPT: &str =
+    "./tests/data/terrain-example_biome.example.example_biome.updated.zsh";
+pub const WITH_NONE_UPDATED_EXAMPLE_TOML: &str = "./tests/data/terrain.example.none.updated.toml";
+pub const WITH_NONE_BIOME_FOR_UPDATED_NONE_EXAMPLE_SCRIPT: &str =
+    "./tests/data/terrain-none.example.none.updated.zsh";
+pub const WITH_EXAMPLE_BIOME_FOR_UPDATED_NONE_EXAMPLE_SCRIPT: &str =
+    "./tests/data/terrain-example_biome.example.none.updated.zsh";
+pub const WITH_AUTO_APPLY_ENABLED_EXAMPLE_TOML: &str =
+    "./tests/data/terrain.example.auto_apply.enabled.toml";
